LYNX-738: Initial implementation
